Change project file structure, add index.ts file to the every component folder, move styled components in separate files
