+++ src/App/App.tsx
@@ -1,26 +1,16 @@
 import React from "react";
-import styled from "styled-components";
 import { BrowserRouter, Switch, Route } from "react-router-dom";
 
-import { Navbar } from "./components/Navbar";
-import { Content } from "./components/Content";
-import { shop } from "./pages/shop";
-import { access } from "./pages/access";
-import { help } from "./pages/help";
-import { authentication } from "./pages/authentication";
-import { Header } from "./components/shared/Header";
-import { Footer } from "./components/shared/Footer";
-import { DummyBanner } from "./components/DummyBanner";
-
-const ContentWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  padding-left: 280px;
-  width: calc(100% - 280px);
-  min-height: 100vh;
-  font-size: 1rem;
-  color: #fff;
-`;
+import { Navbar } from "../components/Navbar";
+import { Content } from "../components/Content";
+import { shop } from "../pages/shop";
+import { access } from "../pages/access";
+import { help } from "../pages/help";
+import { authentication } from "../pages/authentication";
+import { Header } from "../components/shared/Header";
+import { Footer } from "../components/shared/Footer";
+import { DummyBanner } from "../components/DummyBanner";
+import { StyledContentWrapper as ContentWrapper } from "./StyledContentWrapper";
 
 const App: React.FC = () => {
   return (
