+++ src/App/StyledContentWrapper.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const StyledContentWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  padding-left: 280px;
+  width: calc(100% - 280px);
+  min-height: 100vh;
+  font-size: 1rem;
+  color: #fff;
+`;
