+++ src/components/DummyBanner/index.ts
@@ -1,0 +1,1 @@
+export * from "./DummyBanner";
