+++ src/components/Navbar/LinkWrapper.tsx
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const LinkWrapper = styled.div`
+  > * {
+    color: #fff;
+    font-size: 1rem;
+    margin-bottom: 10px;
+    text-decoration: none;
+    outline: none;
+    &:hover {
+      color: #e2c9d9;
+      text-decoration: none;
+    }
+  }
+`;
