+++ src/components/Navbar/NavbarWrapper.tsx
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const StyledNavbarWrapper = styled.div`
+  min-width: 280px;
+  height: 100%;
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  position: fixed;
+  left: 0;
+  top: 0;
+  background: #000;
+  box-sizing: border-box;
+`;
