+++ src/components/shared/Footer/Footer.tsx
@@ -1,21 +1,5 @@
 import React from "react";
-import styled from "styled-components";
-
-const FooterWrapper = styled.div`
-  margin-top: auto;
-  min-height: 100px;
-  width: 100%;
-  padding: 20px;
-  box-sizing: border-box;
-  background-image: linear-gradient(
-    to right top,
-    #ff009b,
-    #c01667,
-    #821b3f,
-    #471721,
-    #120606
-  );
-`;
+import { StyledFooterWrapper as FooterWrapper } from "./StyledFooterWrapper";
 
 export const Footer: React.FC = () => {
   return <FooterWrapper>Footer</FooterWrapper>;
