+++ src/components/shared/Footer/StyledFooterWrapper.tsx
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+export const StyledFooterWrapper = styled.div`
+  margin-top: auto;
+  min-height: 100px;
+  width: 100%;
+  padding: 20px;
+  box-sizing: border-box;
+  background-image: linear-gradient(
+    to right top,
+    #ff009b,
+    #c01667,
+    #821b3f,
+    #471721,
+    #120606
+  );
+`;
