+++ src/components/shared/Footer/index.ts
@@ -1,0 +1,1 @@
+export * from "./Footer";
