+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "@atlaskit/css-reset";
-import App from "./App";
+import App from "./App/App";
 
 ReactDOM.render(<App />, document.getElementById("root"));
