+++ src/pages/access/access.tsx
@@ -1,5 +1,5 @@
 import React from "react";
 
-export const Access: React.FC = () => {
+export const access: React.FC = () => {
   return <div>Access Page</div>;
 };
