+++ src/pages/shop/shop.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { DummyContent } from "../../components/DummyContent";
 
-export const Shop: React.FC = () => {
+export const shop: React.FC = () => {
   return (
     <div>
       <DummyContent />
